Add full block validation
The isValid() overload for blocks now does full verification of a block.

Check the validity of a block.

A block is considered valid if:
    - its height is the previous block height + 1
    - its prev_hash is the previous block header's hash
    - the number of transactions in the block are equal to Block.TxsInBlock
    - the merkle root in the header matches the re-built merkle tree root
      based on the included transactions in the block
    - all the the transactions pass validation, which implies:
        - signatures are authentic
        - the inputs spend an output which must be found with the
          findOutput() delegate

Note that checking for transactions which double-spend is the responsibility
of the findOutput() delegate. During validation, whenever this delegate is
called it should also keep track of the used UTXOs, thereby marking
it as a spent output.

See the `findNotDoubleSpent` function in the unittest for an example.