Implement a BitField structure
It's using BitArray as its backing store,
and additionally has serialization support.

This implementation is specific to our use-case:

Note that the number of bits used is not serialized,
to avoid bloating the Block header size when this
BitField is used in the Block. The bit length can
be derived during deserializations as the length
of bytes of the backing store * 8.

In our case, for example, if we have 6 validators
and we deserialize the bitfield as a multiple of 8,
it is Ok for the last two bits to default-initialize
to 0 as this means the validators with those key
indices haven't signed.